refactor: изменить структуру проекта
